analyzer: finally split overload_set.h into sensible pieces.
