+++ analyzer/statements/statement.cpp
@@ -25,8 +25,8 @@
 #include "vapor/analyzer/expressions/import.h"
 #include "vapor/analyzer/function.h"
 #include "vapor/analyzer/statements/declaration.h"
+#include "vapor/analyzer/statements/function_declaration.h"
 #include "vapor/analyzer/statements/if.h"
-#include "vapor/analyzer/statements/overload_set.h"
 #include "vapor/analyzer/statements/return.h"
 #include "vapor/analyzer/symbol.h"
 #include "vapor/parser.h"
